C:/Program Files/Git/seed, /signin, /register tests passed
