+++ backend/globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:51974/"}
+{"mongoUri":"mongodb://127.0.0.1:64836/"}
