+++ backend/tests/routes.test.js
@@ -1,57 +1,16 @@
 import request  from 'supertest';
 import app from '../server.js'
-import express from 'express'
-import mongoose from 'mongoose'
 import seed from '../util/seed.js';
 import User from '../models/userModel.js';
-import {jest} from '@jest/globals';
 import expressAsyncHandler from 'express-async-handler';
-import userRouter from '../routes/userRouter';
-import {generateToken, isAuth, isAdmin} from '../util/util.js';
+import {generateToken} from '../util/util.js';
+import {removeAllCollections, dropAllCollections } from './helpers';
 import bcrypt from 'bcryptjs';
-
-// jest.useFakeTimers()
-// jest.runAllTimers()
 
 let testUser = {
     email: "king@example.com",
     password: "1234"
 };
-
-async function removeAllCollections() {
-    const collections = Object.keys(mongoose.connection.collections);
-    for (const collectionName of collections) {
-        const collection = mongoose.connection.collections[collectionName];
-        await collection.deleteMany();
-    }
-}
-
-async function dropAllCollections() {
-    const collections = Object.keys(mongoose.connection.collections);
-    for (const collectionName of collections) {
-        const collection = mongoose.connection.collections[collectionName];
-        try {
-            await collection.drop();
-        } catch (error) {
-            // Sometimes this error happens, but you can safely ignore it
-            if (error.message === "ns not found") return;
-            // This error occurs when you use it.todo. You can
-            // safely ignore this error too
-            if (error.message.includes("a background operation is currently running"))
-            return;
-            console.log(error.message);
-        }
-    }
-}
-
-// beforeAll(async () => {
-//     await mongoose.connect('mongodb://localhost/crumple_test', {
-//         useNewUrlParser: true,
-//         useUnifiedTopology: true,
-//         useCreateIndex: true,
-//     });
-//     console.log("done in beforeAll")
-// });
 
 beforeEach( async () => {
     try{
@@ -103,7 +62,7 @@ app.post(
                 });
             }
         }
-        res.status(401).send({ message: 'Invalid email or password' });
+        res.status(400).send({ message: 'Invalid email or password' });
     })
 );
 
@@ -156,6 +115,17 @@ describe('User Endpoints', () => {
         expect(response.statusCode).toBe(200);
     });
 
+    it("should POST to '/signin' and respond with a 400 status code", async() => {
+
+        const response = await request(app)
+            .post("/signin")
+            .send({ 
+                email: "NoUser@example.com", 
+                password: '1234' 
+            })
+        expect(response.statusCode).toBe(400);
+    });
+
     it("should POST to '/register' and respond with a 200 status code", async() => {
 
         const response = await request(app)
@@ -168,36 +138,17 @@ describe('User Endpoints', () => {
         expect(response.statusCode).toEqual(200);
     });
 
-//     it("should repsond with status code 400", async() => {
+    it("should POST to '/register' and respond with a 400 status code", async() => {
 
-//         const response = await request(app)
-//             .post('/register')
-//             .send({
-//                 email: "user5@example.com", 
-//                 password: undefined  
-//             });
+        const response = await request(app)
+            .post('/register')
+            .send({
+                email: "NoUser@example.com", 
+                password: undefined  
+            });
 
-//         expect(response.statusCode).toEqual(400);
-//     });
-
-//     it("should repsond with status code 200 and user object", async() => {
-
-//         const id = '61024b128def3a27b8211796';
-//         const response = await request(app)
-//             .get(`/profile/${id}`)
-
-//         expect(response.statusCode).toEqual(200);
-//     });
-
-//     it("should repsond with status code 404 and message", async() => {
-
-//         const id = ''
-//         const response = await request(app)
-//             .get(`/profile/${id}`)
-
-//         expect(response.statusCode).toEqual(404);
-//         // expect(response.statusCode).toEqual(404);
-//     });
+        expect(response.statusCode).toEqual(500);
+    });
 
 });
